Create GMAT_Release_Orbit.script
This GMAT script models a spacecraft released into an orbit defined by the Cartesian elements (X,Y,Z,VX,VY,VZ) = (0,0,20,2380,0,0). This is my understanding of the orbit into which the lander my group is to design will be released, according to page 2 of "L'SPACE Mission Concept Academy Spring 2020: Investigating the Resource  Potential of Polar Water Ice on the Moon".